requirements.txt
django == 4.2.2
psycopg2 == 2.9.6
python-dotenv == 1.0.0
djangorestframework-simplejwt==5.2.2